create HeaderCell
